Version 11
Completed this app for the most part. Only missing correcting the landscape orientation look and the fact that when multiple stocks are found, I can't scroll down to see more options. Tomorrow I try to fix the orientation and I won't worry about the scrolling part. But turn it in by 3 pm. It is due at 5 pm.